Revitalize the Pool object and associated methods
The rule34Py.get_pool() method returns a list of Post ids that are
assigned to a pool on R34. This method works, but fails to retrieve all
the useful information about the pool (it's name, description, etc.).

Implement a new Pool dataclass, that holds this information. Revitalize
the get_pool() method to parse this info from the pool page's HTML and
generate one of these new objects.

Also implement the backend for an additional dataclass (PoolHistoryEvent) where
we can eventually store events from the pool's history record. It is not
currently exposed.

Regenerate the responses cache.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>