rule34Py: correct user-agent and deprecate __headers__
The default user-agent string in the client contains a 'test' value and
is different from what it was previously.

Correct the user-agent string in the client to reflect what it was
previously. Add a test to confirm that the user-agent string is
well-formed and can be changed.

Deprecate the __vars__.__header__ and __useragent__ variables, as they
are no longer used.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>