docs: support docs builds on py3.9
The tomllib module was added to the standard python distribution
starting in version 3.11. Trying to build the docs on versions prior to
that will throw an import error.

Conditionally prefer to use the third-party 'tomli' module for python
runtimes prior to 3.11. Tomli provides a similar interface and
intentional backport support for TOML parsing on old pythons. It is
licensed under MIT.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>