pyproject.toml: upgrade minimum python to 3.9
Python 3.9 is the earliest python version that is still supported by
python upstream. And it is the earliest version that we reliably test on
GH runners (without going to great lengths to manually install an
earlier python runtime).

So upgrade the minimum supported python version in the package metadata
to 3.9.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>