tests: mark tests protected by captcha
Some unit tests rely on methods that are normally protected by captchas.
When running these tests with the mock responses server in place, caches
values from a captcha-cleared session are returned and there is no
issue.

But when trying to run the tests with the mock server disabled, they are
expected to fail.

Mark the tests as being captcha protected with the pytest 'captcha'
mark, so that they can be excluded from test runs as a group.

```
pytest -m 'not captcha'
```

Modify the Mock34 fixture to honor a new environment variable
`R34_MOCK`. When set to False, cleanly disable the mock server.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>